mustafa | midway implementing vouchers preview API
